// src/app/models/addItemRequest.ts
export interface addItemRequest {
    id: string
    name: string
    type: string
    isbn: string
    issn: string
    category: string
    genre: string
    publicationDate: string
    totalPages: number
    language: string
    description: string
    isAvailable: boolean
    publisherId: string
    locationId: string
  }
  